+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/AnnotationHandlerMapping.java
@@ -3,7 +3,6 @@ package nextstep.mvc.tobe;
 import com.google.common.collect.Maps;
 import nextstep.mvc.HandlerMapping;
 import nextstep.mvc.exception.AnnotationScanFailException;
-import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
 import org.reflections.Reflections;
@@ -18,7 +17,6 @@ import java.util.Map;
 import java.util.Set;
 
 public class AnnotationHandlerMapping implements HandlerMapping {
-    private static final Logger logger = LoggerFactory.getLogger(AnnotationHandlerMapping.class);
     private Object[] basePackage;
 
     private Map<HandlerKey, HandlerExecution> handlerExecutions = Maps.newHashMap();
@@ -30,8 +28,43 @@ public class AnnotationHandlerMapping implements HandlerMapping {
     @Override
     public void initialize() {
         ControllerScanner controllerScanner = new ControllerScanner(new Reflections(basePackage));
-        handlerExecutions = controllerScanner.scan();
+        Set<Class<?>> controllers = controllerScanner.scan();
+        try {
+            initializeHandlerExecutions(controllers);
+        } catch (Exception e) {
+            throw new AnnotationScanFailException(e.getCause());
+        }
         // TODO scan services
+    }
+
+    private void initializeHandlerExecutions(Set<Class<?>> controllers) throws InstantiationException, IllegalAccessException, InvocationTargetException, NoSuchMethodException {
+        for (Class<?> controller : controllers) {
+            Object controllerInstance = controller.getDeclaredConstructor().newInstance();
+            RequestMappingScanner requestMappingScanner = new RequestMappingScanner(controller);
+            Method[] methods = requestMappingScanner.scanMethods();
+            initializeHandlerExecution(controllerInstance, methods);
+        }
+    }
+
+    private void initializeHandlerExecution(Object controllerInstance, Method[] methods) {
+        Arrays.stream(methods)
+            .filter(method -> method.isAnnotationPresent(RequestMapping.class))
+            .forEach(method -> addHandlerExecution(controllerInstance, method));
+    }
+
+    private void addHandlerExecution(Object controllerInstance, Method method) {
+        RequestMapping annotation = method.getAnnotation(RequestMapping.class);
+        String url = annotation.value();
+        RequestMethod[] requestMethods = annotation.method();
+
+        addHandlerKey(controllerInstance, method, url, requestMethods);
+    }
+
+    private void addHandlerKey(Object controllerInstance, Method method, String url, RequestMethod[] requestMethods) {
+        for (RequestMethod requestMethod : requestMethods) {
+            HandlerKey key = new HandlerKey(url, requestMethod);
+            handlerExecutions.put(key, (req, res) -> method.invoke(controllerInstance, req, res));
+        }
     }
 
     @Override
