+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/ArgumentResolver.java
@@ -1,0 +1,4 @@
+package nextstep.mvc.tobe;
+
+public class ArgumentResolver {
+}
