+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/ArgumentResolverTest.java
@@ -1,0 +1,4 @@
+package nextstep.mvc.tobe;
+
+public class ArgumentResolverTest {
+}
