+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/MyControllerTests.java
@@ -1,0 +1,4 @@
+package nextstep.mvc.tobe;
+
+public class MyControllerTests {
+}
